tests/wpt/tests/dom/nodes/Element-firstElementChild-namespace.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>firstElementChild with namespaces</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<h1>Test of firstElementChild with namespaces</h1>
<div id="log"></div>
<p id="parentEl">The result of this test is a unknown.</p>
<script>
test(function() {
  var parentEl = document.getElementById("parentEl")
  var el = document.createElementNS("http://ns.example.org/pickle", "pickle:dill")
  el.setAttribute("id", "first_element_child")
  parentEl.appendChild(el)
  var fec = parentEl.firstElementChild
  assert_true(!!fec)
  assert_equals(fec.nodeType, 1)
  assert_equals(fec.getAttribute("id"), "first_element_child")
  assert_equals(fec.localName, "dill")
})
</script>
